refactor: サーバーコンポーネントでラップ
